+++ src/main/java/pl/pomykalskimateusz/evolution/api/game/play/PlayGameResponse.java
@@ -2,7 +2,9 @@ package pl.pomykalskimateusz.evolution.api.game.play;
 
 import pl.pomykalskimateusz.evolution.domain.model.UserBalance;
 
-public record PlayGameResponse(double balance) {
+import java.math.BigDecimal;
+
+public record PlayGameResponse(BigDecimal balance) {
     public static PlayGameResponse of(UserBalance userBalance) {
         return new PlayGameResponse(userBalance.balance());
     }
