+++ src/main/java/pl/pomykalskimateusz/evolution/api/user/create/CreateUserResponse.java
@@ -2,7 +2,9 @@ package pl.pomykalskimateusz.evolution.api.user.create;
 
 import pl.pomykalskimateusz.evolution.repository.user.UserEntity;
 
-public record CreateUserResponse(Long id, double balance) {
+import java.math.BigDecimal;
+
+public record CreateUserResponse(Long id, BigDecimal balance) {
     public static CreateUserResponse of(UserEntity userEntity) {
         return new CreateUserResponse(userEntity.getId(), userEntity.getBalance());
     }
