+++ src/main/java/pl/pomykalskimateusz/evolution/service/user/UserService.java
@@ -5,9 +5,11 @@ import pl.pomykalskimateusz.evolution.domain.exception.UserNotFound;
 import pl.pomykalskimateusz.evolution.repository.user.UserEntity;
 import pl.pomykalskimateusz.evolution.repository.user.UserRepository;
 
+import java.math.BigDecimal;
+
 @Service
 public record UserService(UserRepository userRepository) {
-    private static final double initialBalance = 5000;
+    private static final BigDecimal initialBalance = BigDecimal.valueOf(5000.00);
 
     public UserEntity createUser() {
         return userRepository.save(buildUserEntity());
